Fix(CTAN): Don't allow special characters in args.
Special characters(except hyphens, underscores, and dots) cannot be used
in a package name.
If we make a request to `/pkg/<arg>` with arg containing special
characters it redirect to some non-relevant page, e.g. backslashes makes
it go to login page, and :A is route for packages start with an A.
But we allow hyphens, underscores, and dots to be in the args because
package names can contain those.