Fix (find_member): prevent a resource-intensive iteration loop in large servers.
If a member had no nickname, its username would be compared with the entire server member list to check for any equalities. This was repeated for every member in the collection. This has now been removed.

Introduced a hard limit on the number of members in the collection, by default this number is 1000 members.